model_inference_triage_pipeline/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
absl-py==0.7.1            # via tensorflow-gpu
arrow==0.13.2
astor==0.8.0              # via tensorflow-gpu
beautifulsoup4==4.7.1
bert-tensorflow==1.0.1
boto3==1.9.157
botocore==1.12.157
cachetools==3.1.1         # via google-auth
certifi==2019.3.9         # via requests
chardet==3.0.4            # via requests
contractions==0.0.18
cycler==0.10.0            # via matplotlib
daiquiri==1.5.0
dill==0.2.9
docutils==0.14            # via botocore
gast==0.2.2               # via tensorflow-gpu
google-api-core==1.11.1   # via google-cloud-bigquery, google-cloud-core
google-auth==1.6.3        # via google-api-core
google-cloud-bigquery==1.12.1
google-cloud-core==1.0.0  # via google-cloud-bigquery
google-resumable-media==0.3.2  # via google-cloud-bigquery
googleapis-common-protos==1.6.0  # via google-api-core
grpcio==1.21.1            # via tensorboard, tensorflow-gpu
h5py==2.9.0               # via keras, keras-applications
idna==2.8                 # via requests
jmespath==0.9.4           # via boto3, botocore
keras-applications==1.0.7  # via keras, tensorflow-gpu
keras-preprocessing==1.0.9  # via keras, tensorflow-gpu
keras==2.2.4
kiwisolver==1.1.0         # via matplotlib
lxml==4.3.3
markdown==3.1.1           # via tensorboard
matplotlib==3.1.1
numpy==1.16.4
pandas==0.24.2
protobuf==3.8.0           # via google-api-core, google-cloud-bigquery, googleapis-common-protos, tensorboard, tensorflow-gpu, tensorflow-hub
pyasn1-modules==0.2.5     # via google-auth
pyasn1==0.4.5             # via pyasn1-modules, rsa
pyparsing==2.4.2          # via matplotlib
python-dateutil==2.8.0    # via arrow, botocore, matplotlib, pandas
pytz==2019.1              # via google-api-core, pandas
pyyaml==5.1               # via keras
requests==2.22.0          # via google-api-core
rsa==4.0                  # via google-auth
s3transfer==0.2.0         # via boto3
scipy==1.3.0              # via keras
six==1.12.0               # via absl-py, bert-tensorflow, cycler, google-api-core, google-auth, google-resumable-media, grpcio, h5py, keras, keras-preprocessing, protobuf, python-dateutil, tensorboard, tensorflow-gpu, tensorflow-hub
soupsieve==1.9.1          # via beautifulsoup4
tensorboard==1.12.2       # via tensorflow-gpu
tensorflow-gpu==1.12.0
tensorflow-hub==0.6.0
termcolor==1.1.0          # via tensorflow-gpu
tqdm==4.35.0
urllib3==1.25.3           # via botocore, requests
werkzeug==0.15.4          # via tensorboard
wheel==0.33.4             # via tensorboard, tensorflow-gpu

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.2.0        # via google-api-core, kiwisolver, markdown, protobuf
